OPS-3094 Refactored parameters for TARGET_URL to be read from environment
